Slowly moving: setting up new component for post submit form, tweaking some more styling, adding a CTA button, scaffolding wouter functionality
